Checkbox group cleanup
