+++ libs/@hashintel/ds-components/src/components/RadioGroup/radio-group.tsx
@@ -3,8 +3,11 @@ import { useId } from "react";
 
 import { cx } from "@hashintel/ds-helpers/css";
 
+import {
+  getGroupFocusProps,
+  styles,
+} from "../../util/radio-checkbox-group-shared";
 import { Radio } from "../Radio/radio";
-import { styles } from "./radio-group.recipe";
 
 import type { SharedInputProps } from "../../util/form-shared";
 
@@ -62,33 +65,6 @@ export const RadioGroup = <const ValueType extends string>({
   const selectedIndex = items.findIndex((item) => item.value === value);
   const autoFocusIndex = selectedIndex === -1 ? 0 : selectedIndex;
 
-  // Focus events bubble, so moving between options would otherwise fire blur
-  // then focus on the group itself. These only report focus genuinely entering
-  // or leaving the group, i.e. when the related element is outside it.
-  const handleGroupFocus = (event: React.FocusEvent<HTMLDivElement>) => {
-    if (!event.currentTarget.contains(event.relatedTarget as Node | null)) {
-      onFocus?.(event as unknown as React.FocusEvent<HTMLInputElement>);
-    }
-  };
-
-  const handleGroupBlur = (event: React.FocusEvent<HTMLDivElement>) => {
-    if (!event.currentTarget.contains(event.relatedTarget as Node | null)) {
-      onBlur?.(event as unknown as React.FocusEvent<HTMLInputElement>);
-    }
-  };
-
-  // Pressing an option's label would otherwise blur the currently-focused
-  // option out to `<body>` (where `relatedTarget` is null) before `click`
-  // focuses the pressed option — surfacing as the group losing then regaining
-  // focus. Preventing the default press behaviour keeps the current option
-  // focused until `click` moves focus straight to the pressed option, so
-  // `relatedTarget` stays within the group and the transient never happens.
-  const handleGroupMouseDown = (event: React.MouseEvent<HTMLDivElement>) => {
-    if ((event.target as HTMLElement).closest("label")) {
-      event.preventDefault();
-    }
-  };
-
   return (
     <ArkRadioGroup.Root
       value={value}
@@ -101,14 +77,12 @@ export const RadioGroup = <const ValueType extends string>({
       disabled={disabled}
       orientation={layout === "inline" ? "horizontal" : "vertical"}
       ids={{ itemHiddenInput: itemInputId }}
-      onMouseDown={handleGroupMouseDown}
-      onFocus={handleGroupFocus}
-      onBlur={handleGroupBlur}
       data-testid={testId}
       ref={ref as React.Ref<HTMLDivElement>}
       className={cx(styles({ layout }), className)}
       required={required}
       invalid={invalid}
+      {...getGroupFocusProps({ onFocus, onBlur })}
       {...ariaProps}
     >
       {items.map((item, index) => {
